changed station id type from Int to Long

+++ app/src/main/java/xyz/skether/radiline/data/backend/converter/Parsers.kt
@@ -41,7 +41,7 @@ fun XmlPullParser.parseTuneInXml(): TuneInXml {
 fun XmlPullParser.parseStationXml(): StationXml {
     require(START_TAG, ns, "station")
     val station = StationXml(
-        id = getAttributeValue(ns, "id").toInt(),
+        id = getAttributeValue(ns, "id").toLong(),
         name = getAttributeValue(ns, "name"),
         genre = getAttributeValue(ns, "genre"),
         currentTrack = getAttributeValue(ns, "ct"),
